breaking up the button definitions for now
