+++ src/plugins/drawing/drawing-toolbar-stamp-button.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import { TileToolbarButton } from "../../components/toolbar/tile-toolbar-button";
+import { IToolbarButtonComponentProps } from "../../components/toolbar/toolbar-button-manager";
+import EllipseIcon from "./assets/ellipse-icon.svg";
+
+import "./drawing-toolbar.scss";
+
+
+export function StampButton({ name }: IToolbarButtonComponentProps) {
+  const selected = false;
+  const title = "Stamp";
+
+  function handleClick() {
+    console.log("handleClick: " + name);
+  }
+
+  return (
+    <TileToolbarButton name={name} title={title} selected={selected} onClick={handleClick}>
+      <EllipseIcon />
+    </TileToolbarButton>
+  );
+}
+
+
